Updated according feedback from code review
